[refactor] thining central net behavior, moving kademlia parts to net storage
